+++ lib/clin/option_list.rb
@@ -1,0 +1,25 @@
+require 'clin'
+require 'clin/option'
+
+class Clin::OptionList < Clin::Option
+
+  # @see Clin::Option#initialize
+  def initialize(*args)
+    super
+    if flag?
+      self.default = 0
+    else
+      self.default = []
+    end
+  end
+
+  def on(value, out)
+    if flag?
+      out[@name] ||= 0
+      out[@name] += 1
+    else
+      out[@name] ||= []
+      out[@name] << value
+    end
+  end
+end