+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2.0
 Features:
     - Allow unknown options to be ignored and not raise Error(#1)
+    - Added list options. For options that can be multiple times in the same command(#2)
+    - Added default value for options.(#3)
 ## 0.1.0
 Inital release.